+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## TBD
+
+IMPROVEMENTS:
+
+- [Query] Filter proxy node is not active `GetIdpNodesInfo`.
+
 ## 0.9.0 (October 4, 2018)
 
 BREAKING CHANGES: